backend/requirements.txt
# backend/requirements.txt
# HuntAI - AI Job Hunter Agent
# Core dependencies (Optimized for SeleniumBase)

fastapi
uvicorn
google-auth
google-auth-oauthlib
openpyxl
pydantic
pydantic-settings
pydantic[email]
httpx==0.24.1
python-multipart
slowapi
pdfplumber
python-docx
motor
dnspython
python-dotenv
groq
seleniumbase
python-jose[cryptography]
passlib[bcrypt]
authlib
email-validator
psutil
